requirements.txt
py-clob-client>=0.1.0
python-dotenv>=1.0.0
colorama>=0.4.6
websocket-client>=1.6.0
requests>=2.31.0
